feat: extract sidebar (DevHub) and cleanup config
